Added DI for PostRepository
