+++ conf/application.conf
@@ -9,3 +9,4 @@ play {
   }
 }
 
+play.modules.enabled += "modules.PostBindings"
